+++ projeto-volt/src/main/java/sptech/school/projetovolt/entity/produto/dto/ProdutoAlteracaoDto.java
@@ -20,5 +20,5 @@ public class ProdutoAlteracaoDto {
     @Size(max = 45)
     private String estadoGeral;
     @PositiveOrZero
-    private Double desconto;
+    private Integer desconto;
 }
